feat: session expiration
This aligns the timeout from within the Relying Party to align with the settings configured on the Authenticator side.
