+++ webauthn/const.go
@@ -1,0 +1,8 @@
+package webauthn
+
+import (
+	"time"
+)
+
+const defaultTimeoutUVD = time.Millisecond * 120000
+const defaultTimeoutUV = time.Millisecond * 300000
